+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/feature/channel/list/ChannelsActivity.kt
@@ -175,7 +175,6 @@ class ChannelsActivity : BaseConnectedActivity() {
             },
             onHeaderActionClick = ::openAddChannel,
         )
-
 //                MyCustomUiSimplified()
 //                MyCustomUi()
     }